requirements.txt
# AtendentePro - Sistema de Atendimento Inteligente com Múltiplos Agentes IA
# Dependências principais do projeto

# OpenAI Agents SDK (core framework)
openai-agents

# Data validation and serialization
pydantic>=2.0.0

# YAML configuration files
PyYAML>=6.0

# Document processing
PyPDF2>=3.0.0
python-docx>=0.8.11
python-pptx>=0.6.21
PyMuPDF>=1.23.0  # fitz

# Environment management
python-dotenv>=1.0.0

# HTTP client (telemetria Application Insights)
httpx>=0.27.0

# Machine Learning and Vector Operations
numpy>=1.24.0
scikit-learn>=1.3.0

# Testing framework
pytest>=7.0.0

# Async support (built-in with Python 3.7+)
# asyncio is part of standard library

# Standard library modules (no installation needed):
# - argparse
# - pathlib
# - typing
# - functools
# - logging
# - json
# - os
# - sys
# - pickle
